Move column-merge pattern utils to _cols_merge.py
